+++ Insurance-System/src/main/java/org/example/customerInfoManageTeam/ComplaintManagementTeam.java
@@ -1,4 +1,6 @@
-package org.example;
+package org.example.customerInfoManageTeam;
+
+import org.example.Complaint;
 
 /**
  * @author USER
