+++ Insurance-System/src/main/java/org/example/customerInfoManageTeam/Counselee.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.example.customerInfoManageTeam;
 
 /**
  * @author USER
