+++ Insurance-System/src/main/java/org/example/customerInfoManageTeam/CounseleeList.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.example.customerInfoManageTeam;
 
 /**
  * @author USER
