+++ Insurance-System/src/main/java/org/example/customerInfoManageTeam/CustomerInfoManageTeam.java
@@ -1,4 +1,7 @@
-package org.example;
+package org.example.customerInfoManageTeam;
+
+import org.example.customerInfoManageTeam.Counselee;
+import org.example.customerInfoManageTeam.CounseleeListImpl;
 
 /**
  * @author USER
