+++ Insurance-System/src/main/java/org/example/rewardSupportTeam/AccidentList.java
@@ -1,4 +1,4 @@
-package org.example.domain.team;
+package org.example.rewardSupportTeam;
 
 
 /**
