+++ Insurance-System/src/main/java/org/example/rewardSupportTeam/AccidentListImpl.java
@@ -1,4 +1,4 @@
-package org.example.domain.team;
+package org.example.rewardSupportTeam;
 
 
 import java.util.ArrayList;
