+++ Insurance-System/src/main/java/org/example/rewardSupportTeam/litigationInfo.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.example.rewardSupportTeam;
 
 /**
  * @author USER
