+++ Insurance-System/src/main/java/org/example/rewardSupportTeam/litigationInfoListImpl.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.example.rewardSupportTeam;
 
 /**
  * @author USER
