+++ Insurance-System/src/main/java/org/example/salesChannelTeam/MeetingCounsel.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.example.salesChannelTeam;
 
 /**
  * @author USER
