+++ Insurance-System/src/main/java/org/example/salesChannelTeam/MeetingCounselListImpl.java
@@ -1,13 +1,13 @@
-package org.example;
+package org.example.salesChannelTeam;
 /**
  * @author USER
  * @version 1.0
  */
-public class BusinessActivityListImpl implements BusinessActivityList {
+public class MeetingCounselListImpl implements MeetingCounselList {
 
-	public BusinessActivity m_BusinessActivity;
+	public MeetingCounsel m_MeetingCounsel;
 
-	public BusinessActivityListImpl(){
+	public MeetingCounselListImpl(){
 
 	}
