+++ Insurance-System/src/main/java/org/example/salesChannelTeam/RequestMeetingInfo.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.example.salesChannelTeam;
 
 /**
  * @author USER
