+++ Insurance-System/src/main/java/org/example/user/CustomerListImpl.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.example.user;
 /**
  * @author USER
  * @version 1.0
