+++ Insurance-System/src/main/java/org/example/user/IncomeInfo.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.example.user;
 
 /**
  * @author USER
